+++ cmd/serve.go
@@ -31,13 +31,15 @@ import (
 	"github.com/uniwise/parrot/internal/cache"
 	"github.com/uniwise/parrot/internal/metrics"
 	"github.com/uniwise/parrot/internal/project"
-	"github.com/uniwise/parrot/internal/rest"
+	publicRest "github.com/uniwise/parrot/internal/rest/v1/public"
+	privateRest "github.com/uniwise/parrot/internal/rest/v1/private"
 	"github.com/uniwise/parrot/internal/storage"
 	"github.com/uniwise/parrot/pkg/poedit"
 )
 
 const (
-	confServerPort  = "server.port"
+	confServerPortPublic  = "server.port.public"
+	confServerPortPrivate = "server.port.private"
 	confServerGrace = "server.gracePeriod"
 
 	confLogLevel  = "log.level"
@@ -98,17 +100,31 @@ by caching exports from poeditor`,
 
 		svc := project.NewService(cli, storageService, cacheInstance, viper.GetDuration(confCacheRenewalThreshold), logrus.NewEntry(logger))
 
-		server, err := rest.NewServer(logrus.NewEntry(logger), svc, viper.GetBool(confPrometheusEnabled))
+		publicServer, err := publicRest.NewServer(logrus.NewEntry(logger), svc)
 		if err != nil {
-			logger.WithError(err).Fatal("Could not instantiate server")
+			logger.WithError(err).Fatal("Could not instantiate public server")
 		}
 
-		port := viper.GetInt(confServerPort)
+		publicPort := viper.GetInt(confServerPortPublic)
 
-		logger.Infof("Server listening at :%d", port)
+		logger.Infof("Public server listening at :%d", publicPort)
 		go func() {
-			if err := server.Start(port); err != nil && !errors.Is(err, http.ErrServerClosed) {
-				logger.Fatal("shutting down server")
+			if err := publicServer.Start(publicPort); err != nil && !errors.Is(err, http.ErrServerClosed) {
+				logger.Fatal("shutting down public server")
+			}
+		}()
+
+		privateServer, err := privateRest.NewServer(logrus.NewEntry(logger), svc, viper.GetBool(confPrometheusEnabled))
+		if err != nil {
+			logger.WithError(err).Fatal("Could not instantiate private server")
+		}
+
+		privatePort := viper.GetInt(confServerPortPrivate)
+
+		logger.Infof("Private server listening at :%d", privatePort)
+		go func() {
+			if err := privateServer.Start(privatePort); err != nil && !errors.Is(err, http.ErrServerClosed) {
+				logger.Fatal("shutting down private server")
 			}
 		}()
 
@@ -124,7 +140,10 @@ by caching exports from poeditor`,
 		<-quit
 		ctx, cancel := context.WithTimeout(context.Background(), viper.GetDuration(confServerGrace))
 		defer cancel()
-		if err := server.Shutdown(ctx); err != nil {
+		if err := privateServer.Shutdown(ctx); err != nil {
+			logger.Fatal(err)
+		}
+		if err := publicServer.Shutdown(ctx); err != nil {
 			logger.Fatal(err)
 		}
 	},
@@ -137,7 +156,8 @@ func init() {
 		cDir = "/tmp"
 	}
 
-	viper.SetDefault(confServerPort, 80)
+	viper.SetDefault(confServerPortPublic, 80)
+	viper.SetDefault(confServerPortPrivate, 81)
 	viper.SetDefault(confServerGrace, time.Second*10)
 
 	viper.SetDefault(confLogLevel, "info")
