+++ internal/rest/v1/private/controllers/project.go
@@ -1,107 +1,13 @@
-package v1
+package controllers
 
 import (
-	"bytes"
-	"context"
-	"errors"
-	"fmt"
 	"net/http"
 	"time"
 
 	"github.com/labstack/echo/v4"
 	"github.com/sirupsen/logrus"
 	"github.com/uniwise/parrot/internal/project"
-	"github.com/uniwise/parrot/pkg/poedit"
 )
-
-type getProjectLanguageRequest struct {
-	Project  int    `param:"project" validate:"required"`
-	Language string `param:"language" validate:"required,languageCode"`
-	Format   string `query:"format" validate:"omitempty,oneof=po pot mo xls xlsx csv ini resw resx android_strings apple_strings xliff properties key_value_json json yml xlf xmb xtb arb rise_360_xliff"`
-	Version  string `query:"version" validate:"omitempty,max=20,alphanum"`
-}
-
-func (h *Handlers) getProjectLanguage(ctx echo.Context, l *logrus.Entry) error {
-	req := new(getProjectLanguageRequest)
-	if err := ctx.Bind(req); err != nil {
-		l.WithError(err).Error("Error binding request")
-
-		return echo.ErrBadRequest
-	}
-
-	l = l.WithFields(logrus.Fields{
-		"project":  req.Project,
-		"language": req.Language,
-		"format":   req.Format,
-		"version":  req.Version,
-	})
-
-	if err := ctx.Validate(req); err != nil {
-		l.WithError(err).Error("Error validating request")
-
-		return echo.ErrBadRequest
-	}
-
-	format := "key_value_json"
-	if req.Format != "" {
-		format = req.Format
-	}
-
-	version := "latest"
-	if req.Version != "" {
-		version = req.Version
-	}
-
-	contentMeta, err := poedit.GetContentMeta(format)
-	if err != nil {
-		l.WithError(err).Errorf("No content meta found for format %s", format)
-
-		return echo.ErrBadRequest
-	}
-
-	trans, err := h.ProjectService.GetTranslation(
-		ctx.Request().Context(),
-		req.Project,
-		req.Language,
-		format,
-		version,
-	)
-
-	if errors.Is(err, context.Canceled) {
-		return echo.NewHTTPError(499, "client closed request")
-	}
-
-	if err != nil {
-		switch err.(type) {
-		case *poedit.ErrProjectPermissionDenied:
-			return echo.ErrBadRequest
-		case *poedit.ErrLanguageNotFound,
-			*project.ErrLanguageNotFoundInStorage:
-			return echo.ErrNotFound
-		default:
-			l.WithError(err).Error("Error retrieving translation")
-
-			return echo.ErrInternalServerError
-		}
-	}
-
-	if ctx.Request() == nil {
-		l.Error("Request is nil")
-
-		return errors.New("request is nil")
-	}
-
-	if ctx.Request().Header.Get("If-None-Match") == trans.Checksum {
-		return ctx.NoContent(http.StatusNotModified)
-	}
-
-	ctx.Response().Header().Add("Etag", trans.Checksum)
-	ctx.Response().Header().Add("Cache-Control", fmt.Sprintf("max-age=%.0f", trans.TTL.Seconds()))
-	ctx.Response().Header().Add("Content-Disposition", fmt.Sprintf("filename=%d-%s.%s", req.Project, req.Language, contentMeta.Extension))
-	ctx.Response().Header().Add("Content-Transfer-Encoding", "8bit")
-
-	return ctx.Stream(http.StatusOK, contentMeta.Type, bytes.NewReader(trans.Data))
-}
 
 type getProjectItemResponse struct {
 	ID               int64  `json:"id"`
